Watch BUILD files for changes to refresh CodeLenses.
In addition to updating on file changes, we clear out the lenses when a
file is being edited (i.e., dirty). This is because the lenses don't
automatically move with edits (a more robust language server would be
needed to handle this), and we can't accurately determine the build
targets in an edited BUILD file until it's been saved and we can run the
query again.