Don't fail for "acceptable" buildifier exits.
In some cases, buildifier exits with a non-zero exit code in
situations where we still want to consider that "successful"
for the purposes of the IDE integration; i.e., it correctly
finds errors or warnings in a BUILD/.bzl file. In those cases,
we need to treat the case as a success instead of rejecting
the promise and causing an uncaught error to bubble up.