Fix configuration property scope.
The configuration properties `bazel.executable` and `bazel.buildifierExecutable` are used by the plug-in to determine the path of an executable to run when building / testing targets or when linting / formatting *.bzl files.

As they did not have an associated scope, it was possible to override them in the configuration file of a workspace. This means that a malicious project could cause an arbitrary executable to run when opening a *.bzl file.

This patch sets the configuration scope to `machine`, which means it can be set in user or remote settings, but no longer in the workspace-specific configuration.