Make queries share the same server by default (#270)
The bazel query used to introspect the targets available incorrectly traverses convenience symlinks when the output base has changed. This is a bug in Bazel (see https://github.com/bazelbuild/bazel/issues/10653). A lot of people are hitting this, since it happens with the default setup of bazel and this plugin. Until it is fixed upstream, this works around this issue by enabling the `queriesShareServer` option by default.

Workaround for https://github.com/bazelbuild/vscode-bazel/issues/216

Co-authored-by: Cameron Martin <cameronmartin123@gmail.com>