Add support for launching an LSP (#317)
This adds an option for launching [facebook's starlark LSP server](https://github.com/facebookexperimental/starlark-rust). Currently this has to be built separately and is very experimental since the LSP server seems more of a proof-of-concept. This disables some functionality of the extension (go to definition and the symbol provider), since it is provided by the LSP instead.