Enable the "Restart LSP" command only if a LSP is configured
This commit hides the "Restart LSP" command from the Command Palette if
the LSP is disabled. Otherwise, trying to use the "Restart LSP" command
without any LSP being configured leads to a user-visible error.

As part of this, I also renamed `vscodeBazelHaveBazelWorkspace` to
`bazel.haveWorkspace` such that the context variables all share a common
`bazel.*` prefix.