Activate the extension automatically (#344)
So far, the `bazelWorkspace` view, listing the available build targets,
was only shown after running the first Bazel command, e.g., through
the command pallete. This behavior was probably unintended and is rather
cumbersome for people which want to inspect the available Bazel targets.

The chain of events which lead to this behavior are:
* The `bazelWorkspace` view is only shown when the
  `vscodeBazelHaveBazelWorkspace` context variable is set.
* This context variable is unset by default, and hence the view is
  hidden initially.
* As soon as the first Bazel command is run from the command pallete,
  the extension was activated.
* As part of activating, the extension detected that there is in fact
  a Bazel workspace and hence it set the `vscodeBazelHaveBazelWorkspace`
  context variable, unhiding the `bazelWorkspace` view.

This commit fixes the issue by activating the extension as soon as one
of Bazel's build files is found.