Stricter typescript configuration (#358)
This commit makes the TypeScript configuration stricter by enabling the
`strict` setting. This implies that the `alwaysStrict` setting is
implicitely enabled and no longer needs to be mentioned explicitly.

This required three fixes:
* in `icons.ts`, `SPECIFIC_RULE_CLASS_ICONS` needed an explicit type
* in `bazel_query.ts`, `execOptions` needed an explicit type
* All usages of `blaze_query.Target` had to be replaced by `ITarget`.
  This was necessary, because `queryResult.target` is of type
  `ITarget[]` not `Target` and `bazel_{package,target}_tree_item.ts`
  would not have compiled otherwise.

The `strictNullChecks` setting is explicitly opted-out. There are 117
violations against those stricter null checks, and to make the reviews
easier, I plan to only enable fix those in a separate commit.

I also removed the `allowJs` setting. Not sure why this was ever set. It
doesn't seem to be required anymore.