build: Update toolchain (#363)
This commit updates typescript, eslint and the eslint plugins to the
latest versions.

Also, this commit migrates from `.eslintrc.js` to `eslint.config.js`.
`.eslintrc.js` will be deprecated in ESLint 9.0. This commit migrates
to `eslint.config.js` in anticipation of this change. Doing so also
has the benefit that the config is more self-contained now and no longer
depends on parameters passed in from the `package.json`.