feat(coverage): Demangling C++ & Rust function names in coverage results
Demangle C++ and Rust symbols using `c++filt` and `rustfilt`. We rely on
those tools to be installed on the $PATH.

I was also considering whether I should compile `c++filt` and / or
`rustfilt` to WebAssembly, but decided against it, given that this would
make the build much more complicated and it is currently unclear to me
whether we will stick with the "traditional" tool chains (npm, webpack,
cargo, ...) or if we will pivot to using Bazel.