Use yaml rather than json for bazelrc grammar
YAML is easier to work with for a number of reasons, especially because
(depending on string style used) you don't need to json-escape strings,
which is especially helpful for regexes.  It also tends to
version-control better, and permits adding comments.

This requires promoting js-yaml (which was already a transitive
dependency) to a direct dev dependency, and adding a line to build.sh to
do the conversion.