ci: Build the VSCode extension using GitHub Actions (#364)
This commit adds a Github action which builds the VS Code extension.
The .vsix file is uploaded as an artifact such that it can be easily
downloaded from GitHub (in the "Artifacts" section of the action's
summary page). Those artifacts are also uploaded as part of pull
requests, which makes it easier to quickly download, install and test
the extension built from a pull request.

In addition, if the workflow is triggered because a new release was
tagged, the .vsix is also attached as an artifact to the newly created
release.

As part of this change, the `vsce` tool is now listed as a devDependency
and thereby its exact version is now part of the `package-lock.json`,
making the packaging step more reproducible. The new, recommend way to
package the extension is now `npm run package`.

Co-authored-by: Adam Azarchs <adam.azarchs@10xgenomics.com>
Co-authored-by: Cameron Martin <cameronmartin123@gmail.com>