fix: Default to Python syntax highlighting
This fixes #396. Changes the `.starlark` name suffix to `.python.starlark`.

By using both `python` and `starlark`, vscode is able to default to Python
highlighting if a color theme defines colors specifically for `python`.
For example, default color themes in vscode use
`keyword.operator.logical.python` to highlight `and`/`or`/`not`/`in`.