fix(coverage): Expand external local repo in coverage report
Rules under repo local_repository() are external repo but symlinked
to a local path, usually in the same workspace.

No matter if the local path is in the same workspace, the users should
usually want to check the local path specified in local_repository rules
instead of the path in bazel output
(<bazel workspace>/bazel-<repo>/external/...) in
file explorer.

Works towards #416. Imporves #362.