refactor(workspace-tree): Future proof icon path resolution for minified pakcage
Current resource path resolution is using the source file as the base
directory, this would fail if we minify the package into single js.

Use extension path as the base directory is more reliable: it is also
working for the current packaging, and we just have to make sure the
resource layout of the minified package remains the same.

An enum is provided with a test looping through all value and checking
the existance of the icon files in extension.