refactor(workspace-tree): Future proof icon path resolution (#424)
Current resource path resolution is using the source file as the base
directory. This will fail if we minify the package into single js.

With this commit, we use the extension path as the base directory
which is more reliable: it is also working for the current packaging,
and we just have to make sure the resource layout of the minified
package remains the same.

An enum is provided with a test looping through all value and checking
the existence of the icon files in extension.