feat(quickpick): Introduce dynamic quickpick for targets/packages.
Dynamic quickpick will run bazel queries in the background as the user types. This will limit the query for relevant labels to just a subset of the repository, which can greatly improve feedback time in case of hughe bazel workspaces. The default behaviour is to start with querying the complete workspace (as before) but to cancel a running query as soon as the user starts typing and to start a new query with the given input.