feat(navigation): Add 'Copy Label To Clipboard' to command palette (#467)
This PR adds the already existing (but hidden) command to the command palette to make it usable interactively.

To use it, place your cursor onto a label string inside an open text file.
If the string can be interpreted as a bazel label, the full label is formed and copied to the clipboard.

Part of #194