fix(buildifier): Support formatting module.bazel files
The logic for computing the file type to pass to buildifier was not updated to support module.bazel files. However, support for inferring the file type based on the passed-in file path was added to buildifier back in 2019, so I've updated it to use this.

I also added a test to ensure the diagnostics are being added correctly.

Fixes #471