fix(buildifier): Support linting & formatting MODULE.bazel files (#487)
This bug was caused by the logic for computing the file type to pass to buildifier not supporting module.bazel files. However, support for inferring the file type based on the passed-in file path was added to buildifier back in 2019, so I've updated it to use this.

I also added a test to ensure the diagnostics are being added correctly. This requires buildifier to be installed on CI so I've added an action to install this too.

Fixes #471